test: split Claude conductor helpers
