+++ tests/functional/workers/inference/claude/conductor_scenarios_test.go
@@ -1,0 +1,207 @@
+package claude
+
+import (
+	"context"
+	"errors"
+	"os"
+	"strings"
+	"testing"
+	"time"
+
+	factorysessions "github.com/portpowered/infinite-you/pkg/services/factory_sessions"
+	"github.com/portpowered/infinite-you/tests/functional/internal/support"
+)
+
+func (fixture *claudeDefaultLaneFixture) runScenario(t *testing.T, scenario claudeScenario) {
+	t.Helper()
+
+	opened := support.OpenFactorySessionAt(t, fixture.baseURL, scenario.factoryDir)
+	if opened.Session == nil {
+		t.Fatalf("%s open response missing session: %#v", scenario.name, opened)
+	}
+	sessionID := opened.Session.Id
+	if sessionID == "" || sessionID == factorysessions.DefaultSessionID {
+		t.Fatalf("%s session id = %q, want unique non-default explicit session", scenario.name, sessionID)
+	}
+	fixture.opened.Add(1)
+	closed := false
+	closeSession := func() {
+		if closed {
+			return
+		}
+		support.CloseFactorySessionAt(t, fixture.baseURL, sessionID)
+		closed = true
+		fixture.closed.Add(1)
+	}
+	t.Cleanup(func() {
+		closeSession()
+	})
+	t.Cleanup(scenario.runner.Release)
+
+	responseStream := support.OpenFactoryResponseEventStreamAt(
+		t,
+		support.SessionResponseEventsURL(fixture.baseURL, sessionID),
+	)
+	fixture.streamsOpened.Add(1)
+	// Fast controlled outcomes, especially context.Canceled, can publish their
+	// terminal response event before a just-opened SSE subscription is attached.
+	// Release the external command edge only after the public stream is ready so
+	// the race suite observes the same event boundary on every schedule.
+	scenario.runner.Release()
+	support.WaitForSessionTerminalStatus(t, fixture.baseURL, sessionID, claudeConductorRunTimeout)
+	session := getClaudeSession(t, fixture.baseURL, sessionID)
+	listed := listClaudeSessionWork(t, fixture.baseURL, sessionID)
+	events := support.GetFactoryEventsForSessionAt(t, fixture.baseURL, sessionID)
+	responseEvents := readClaudeResponseEventsUntilTerminal(t, responseStream, claudeConductorRunTimeout)
+
+	assertClaudeWork(t, scenario, listed)
+	dispatchIDs := assertClaudeDispatch(t, scenario, sessionID, events)
+	assertClaudeCommand(t, fixture.router, scenario)
+	providerSessionID := assertClaudeProviderSession(t, scenario, events)
+	assertClaudeEventScope(t, scenario, sessionID, events)
+	responseEventIDs := assertClaudeResponseEvents(t, scenario, sessionID, responseEvents)
+	assertClaudeGoldenScenario(t, scenario, events, responseEvents)
+
+	closeSession()
+	responseStream.WaitClosed(claudeConductorRunTimeout)
+	fixture.streamsClosed.Add(1)
+	assertClaudeSessionDeleted(t, fixture.baseURL, sessionID)
+	fixture.recordObservation(claudeScenarioObservation{
+		sessionID:         session.Id,
+		workID:            scenario.workID,
+		requestID:         scenario.requestID,
+		dispatchIDs:       dispatchIDs,
+		providerSessionID: providerSessionID,
+		responseEventIDs:  responseEventIDs,
+	})
+}
+
+func (fixture *claudeDefaultLaneFixture) assertSharedProcessCleanup(t *testing.T) {
+	t.Helper()
+
+	closeCtx, cancel := context.WithTimeout(context.Background(), claudeConductorRunTimeout)
+	defer cancel()
+	fixture.command.Stop(t)
+	if err := fixture.process.Close(closeCtx); err != nil {
+		t.Fatalf("close shared Claude application process: %v", err)
+	}
+	select {
+	case <-fixture.apiStopped:
+	case <-time.After(claudeConductorRunTimeout):
+		t.Fatal("shared Claude API server did not close after process cleanup")
+	}
+
+	expectedSessions := fixture.opened.Load()
+	if expectedSessions == 0 {
+		t.Fatal("cleanup probe observed no executed Factory Session")
+	}
+	if got := fixture.closed.Load(); got != expectedSessions {
+		t.Fatalf("closed Factory Sessions = %d, want %d", got, expectedSessions)
+	}
+	if got := fixture.streamsOpened.Load(); got != expectedSessions {
+		t.Fatalf("opened response collectors = %d, want %d", got, expectedSessions)
+	}
+	if got := fixture.streamsClosed.Load(); got != expectedSessions {
+		t.Fatalf("closed response collectors = %d, want %d", got, expectedSessions)
+	}
+	for _, scenario := range fixture.scenarios {
+		if got := scenario.runner.ActiveCallCount(); got != 0 {
+			t.Fatalf("%s active Claude command calls after process cleanup = %d, want 0", scenario.name, got)
+		}
+	}
+	removeClaudeOwnedDirectories(t, fixture)
+}
+
+func removeClaudeOwnedDirectories(t *testing.T, fixture *claudeDefaultLaneFixture) {
+	t.Helper()
+
+	ownedDirs := make([]string, 0, len(fixture.scenarios)+1)
+	ownedDirs = append(ownedDirs, fixture.hostDir)
+	for _, scenario := range fixture.scenarios {
+		ownedDirs = append(ownedDirs, scenario.factoryDir)
+	}
+	for _, path := range ownedDirs {
+		if err := os.RemoveAll(path); err != nil {
+			t.Fatalf("remove test-owned Factory directory %q: %v", path, err)
+		}
+		if _, err := os.Stat(path); !errors.Is(err, os.ErrNotExist) {
+			t.Fatalf("test-owned Factory directory %q still exists after cleanup: %v", path, err)
+		}
+	}
+}
+
+func (fixture *claudeDefaultLaneFixture) recordObservation(observation claudeScenarioObservation) {
+	fixture.ledgerMu.Lock()
+	defer fixture.ledgerMu.Unlock()
+	fixture.ledger[observation.requestID] = observation
+}
+
+func (fixture *claudeDefaultLaneFixture) assertSharedIdentityLedger(t *testing.T) {
+	t.Helper()
+
+	fixture.ledgerMu.Lock()
+	observations := make([]claudeScenarioObservation, 0, len(fixture.ledger))
+	for _, observation := range fixture.ledger {
+		observations = append(observations, observation)
+	}
+	fixture.ledgerMu.Unlock()
+	if len(observations) == 0 {
+		t.Fatal("shared-process scenario observations are empty")
+	}
+
+	seenSessions := make(map[string]string, len(observations))
+	seenWorks := make(map[string]string, len(observations))
+	seenRequests := make(map[string]string, len(observations))
+	seenDispatches := make(map[string]string, len(observations))
+	seenProviderSessions := make(map[string]string, len(observations))
+	seenResponseEvents := make(map[string]string)
+	for _, observation := range observations {
+		assertClaudeUniqueIdentity(t, seenSessions, observation.sessionID, observation.requestID, "Factory Session")
+		assertClaudeUniqueIdentity(t, seenWorks, observation.workID, observation.requestID, "Work")
+		assertClaudeUniqueIdentity(t, seenRequests, observation.requestID, observation.requestID, "request")
+		for _, dispatchID := range observation.dispatchIDs {
+			assertClaudeUniqueIdentity(t, seenDispatches, dispatchID, observation.requestID, "dispatch")
+		}
+		if observation.providerSessionID != "" {
+			assertClaudeUniqueIdentity(t, seenProviderSessions, observation.providerSessionID, observation.requestID, "Provider Session")
+		}
+		for _, responseEventID := range observation.responseEventIDs {
+			assertClaudeUniqueIdentity(t, seenResponseEvents, responseEventID, observation.requestID, "response event")
+		}
+	}
+	if len(observations) != len(fixture.scenarios) {
+		// An anchored subtest intentionally exercises only the selected route;
+		// the full parent gate below is the evidence for all four scenarios.
+		return
+	}
+	if got := fixture.opened.Load(); got != int32(len(fixture.scenarios)) {
+		t.Fatalf("Factory Session opens = %d, want %d", got, len(fixture.scenarios))
+	}
+	if got := fixture.closed.Load(); got != int32(len(fixture.scenarios)) {
+		t.Fatalf("Factory Session closes = %d, want %d", got, len(fixture.scenarios))
+	}
+	if got := fixture.identities.sessionCount(); got < uint64(len(fixture.scenarios)) {
+		t.Fatalf("Factory Session IDs generated = %d, want at least %d explicit sessions", got, len(fixture.scenarios))
+	}
+	if got := fixture.apiStarts.Load(); got != 1 {
+		t.Fatalf("API server starts = %d, want exactly one shared process server", got)
+	}
+	wantProviderCalls := 0
+	for _, scenario := range fixture.scenarios {
+		wantProviderCalls += scenario.wantProviderCalls
+	}
+	if got := fixture.router.callCount(); got != wantProviderCalls {
+		t.Fatalf("shared process routed provider calls = %d, want %d", got, wantProviderCalls)
+	}
+}
+
+func assertClaudeUniqueIdentity(t *testing.T, seen map[string]string, value, owner, kind string) {
+	t.Helper()
+	if strings.TrimSpace(value) == "" {
+		t.Fatalf("%s identity for %s is empty", kind, owner)
+	}
+	if previous, ok := seen[value]; ok {
+		t.Fatalf("%s identity %q is shared by %s and %s", kind, value, previous, owner)
+	}
+	seen[value] = owner
+}
